+++ app/components/company/HistoryCard.tsx
